+++ src/operations/slack_handler.py
@@ -1,7 +1,7 @@
 from operations.points import process_match, generate_string
 from operations.leaderboard import generate_leaderboard
 from operations.help import help_text
-from .models import db, SlackTeam
+from ..models import db, SlackTeam
 import re
 import config
