+++ src/strings.json
